requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
Pillow>=9.0.0
langdetect>=1.0.9
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface_hub>=0.16.0
edge-tts>=6.1.9
asyncio>=3.4.3
